feat: extract Reward Modal
