+++ wondrous-bot-admin/src/components/AddFormEntity/components/OptionRewards.tsx
@@ -1,43 +1,31 @@
-import { Box, ButtonBase, Grid } from "@mui/material";
+import AddIcon from "@mui/icons-material/Add";
+import { Box, Grid } from "@mui/material";
 import { PAYMENT_OPTIONS } from "components/CreateTemplate/RewardUtils";
 import DeleteIcon from "components/Icons/Delete";
-import { DiscordRoleIcon, NFTIcon, PointsIcon } from "components/Icons/Rewards";
+import { DiscordRoleIcon, NFTIcon, TokensIcon } from "components/Icons/Rewards";
 import { ButtonIconWrapper } from "components/Shared/styles";
 import { TextLabel } from "components/ViewQuest/styles";
-import AddIcon from "@mui/icons-material/Add";
-import { useMemo } from "react";
 
 const ICONS_MAP = {
   [PAYMENT_OPTIONS.DISCORD_ROLE]: DiscordRoleIcon,
   [PAYMENT_OPTIONS.POAP]: NFTIcon,
-  [PAYMENT_OPTIONS.TOKEN]: PointsIcon,
+  [PAYMENT_OPTIONS.TOKEN]: TokensIcon,
 };
 
-const RewardContent = ({ reward, handleAddReward }) => {
+const RewardContent = ({ reward }) => {
   const Icon = ICONS_MAP[reward?.type];
-
-  const label = useMemo(() => {
-    if (!reward) return null;
-    if (reward?.type === null) {
-      return `Select reward`;
-    }
-    if (reward?.type === PAYMENT_OPTIONS.DISCORD_ROLE) {
-      return `Role: ${reward?.discordRewardData?.discordRoleName}`;
-    }
-    if (reward?.type === PAYMENT_OPTIONS.TOKEN) {
-      return `Token: ${reward?.amount} ${reward?.paymentMethod?.name}`;
-    }
-    if (reward?.type === PAYMENT_OPTIONS.POAP) {
-      return `POAP: ${reward?.poapRewardData?.name}`;
-    }
-  }, [reward]);
+  const label = {
+    [PAYMENT_OPTIONS.DISCORD_ROLE]: `Role: ${reward?.discordRewardData?.discordRoleName}`,
+    [PAYMENT_OPTIONS.TOKEN]: `Token: ${reward?.amount} ${reward?.paymentMethod?.name}`,
+    [PAYMENT_OPTIONS.POAP]: `POAP: ${reward?.poapRewardData?.name}`,
+  };
 
   return (
     <>
       {Icon ? <Icon /> : null}
 
       <TextLabel fontSize="14px" fontWeight="400">
-        {label}
+        {label[reward?.type]}
       </TextLabel>
     </>
   );
@@ -67,8 +55,12 @@ const OptionRewards = ({ rewards, handleRewardDelete, handleAddReward, id }) => 
               flex="1"
               gap="8px"
               onClick={() => addEmptyReward(reward)}
+              color="#000"
+              sx={{
+                cursor: "pointer",
+              }}
             >
-              <RewardContent reward={reward} handleAddReward={() => handleAddReward(id)} />
+              {reward?.type === null ? "Select a reward" : <RewardContent reward={reward} />}
             </Box>
             {rewards?.length > 1 ? (
               <ButtonIconWrapper onClick={() => handleRewardDelete(idx)}>
