+++ wondrous-bot-admin/src/components/CreateTemplate/RewardComponent.tsx
@@ -1,105 +1,14 @@
-import { useLazyQuery, useMutation } from "@apollo/client";
 import { Divider, Grid } from "@mui/material";
-import { useTour } from "@reactour/tour";
 import {
   PAYMENT_OPTIONS,
-  RewardFooterLeftComponent,
-  RewardMethod,
-  RewardMethodOptionButton,
   RewardWrapper,
   RewardWrapperWithTextField,
   RewardsComponent,
 } from "components/CreateTemplate/RewardUtils";
-import { DiscordRoleIcon, PoapIcon, PointsIcon, TokensIcon } from "components/Icons/Rewards";
-import { PricingOptionsTitle, getPlan } from "components/Pricing/PricingOptionsListItem";
-import Modal from "components/Shared/Modal";
+import { DiscordRoleIcon, PointsIcon, TokensIcon } from "components/Icons/Rewards";
 import { SharedSecondaryButton } from "components/Shared/styles";
-import { CREATE_CMTY_PAYMENT_METHOD } from "graphql/mutations/payment";
-import { GET_ORG_DISCORD_ROLES } from "graphql/queries/discord";
-import { GET_CMTY_PAYMENT_METHODS_FOR_ORG } from "graphql/queries/payment";
-import { useContext, useEffect, useState } from "react";
-import truncateEthAddress from "truncate-eth-address";
-import GlobalContext from "utils/context/GlobalContext";
-import { usePaywall, useSubscription } from "utils/hooks";
-import { Label, PoapImage } from "./styles";
-
-const useSubscriptionPaywall = () => {
-  const subscription = useSubscription();
-  const plan = getPlan(subscription?.tier);
-  const { setPaywall, setPaywallMessage } = usePaywall();
-  return {
-    plan,
-    setPaywall,
-    setPaywallMessage,
-  };
-};
-
-const useDiscordRoleRewardData = () => {
-  const { activeOrg } = useContext(GlobalContext);
-  const [getCmtyOrgDiscordRoles, { data: getCmtyOrgDiscordRolesData, variables }] = useLazyQuery(
-    GET_ORG_DISCORD_ROLES,
-    {
-      fetchPolicy: "cache-and-network",
-    }
-  );
-  useEffect(() => {
-    if (activeOrg?.id) {
-      getCmtyOrgDiscordRoles({
-        variables: {
-          orgId: activeOrg?.id,
-        },
-      });
-    }
-  }, [activeOrg?.id]);
-  const discordRoleData = getCmtyOrgDiscordRolesData?.getCmtyOrgDiscordRoles || [];
-  const discordRoles =
-    getCmtyOrgDiscordRolesData?.getCmtyOrgDiscordRoles?.length > 0
-      ? getCmtyOrgDiscordRolesData?.getCmtyOrgDiscordRoles[0]?.roles
-      : [];
-  const discordRoleOptions = discordRoles?.map((role) => ({
-    label: role.name,
-    value: role.id,
-  }));
-
-  return {
-    discordRoleOptions,
-    discordRoleData,
-  };
-};
-
-const useTokenRewardData = () => {
-  const { activeOrg } = useContext(GlobalContext);
-
-  const [createPaymentMethod] = useMutation(CREATE_CMTY_PAYMENT_METHOD, {
-    refetchQueries: [GET_CMTY_PAYMENT_METHODS_FOR_ORG],
-  });
-
-  const [getCmtyPaymentMethods, { data: getCmtyPaymentMethodsData }] = useLazyQuery(GET_CMTY_PAYMENT_METHODS_FOR_ORG, {
-    fetchPolicy: "cache-and-network",
-  });
-
-  const paymentMethods = getCmtyPaymentMethodsData?.getCmtyPaymentMethodsForOrg || [];
-  const paymentMethodOptions = paymentMethods?.map((method) => ({
-    label: method?.name || truncateEthAddress(method?.contractAddress),
-    value: method?.id,
-  }));
-
-  useEffect(() => {
-    if (activeOrg?.id) {
-      getCmtyPaymentMethods({
-        variables: {
-          orgId: activeOrg?.id,
-        },
-      });
-    }
-  }, [activeOrg?.id]);
-
-  return {
-    paymentMethodOptions,
-    paymentMethods,
-    createPaymentMethod,
-  };
-};
+import RewardModal, { useAddRewardModalState } from "./RewardModal";
+import { PoapImage } from "./styles";
 
 const handleDiscordRoleRewardRemove = ({ reward, setQuestSettings }) => {
   setQuestSettings((prev) => {
@@ -116,23 +25,6 @@ const handleDiscordRoleRewardRemove = ({ reward, setQuestSettings }) => {
   });
 };
 
-const handleNewDiscordRole = ({ newReward, rewards, onRewardAdd, discordRoleOptions, discordRoleData }) => {
-  const discordRoleSelected = discordRoleOptions.find((option) => option.value === newReward.value);
-  const discordRoleAlreadyExists = rewards.some(
-    (reward) => reward.type === "discord_role" && reward.discordRewardData.discordRoleId === newReward?.value
-  );
-  if (!discordRoleAlreadyExists) {
-    onRewardAdd({
-      type: newReward?.type,
-      discordRewardData: {
-        discordRoleId: discordRoleSelected?.value,
-        discordGuildId: discordRoleData[0]?.guildId,
-        discordRoleName: discordRoleSelected?.label,
-      },
-    });
-  }
-};
-
 const OnPaymentMethodRewardRemove = ({ reward, setQuestSettings }) => {
   setQuestSettings((prev) => {
     const newRewards = prev.rewards.filter((r) => {
@@ -146,120 +38,6 @@ const OnPaymentMethodRewardRemove = ({ reward, setQuestSettings }) => {
       rewards: newRewards,
     };
   });
-};
-
-const handleAddTokenOnModal = ({
-  newReward,
-  rewards,
-  onRewardAdd,
-  paymentMethod,
-  setErrors,
-  errors,
-  setQuestSettings,
-  handleToggle,
-  addPaymentMethod,
-  createPaymentMethod,
-  activeOrg,
-  setAddPaymentMethod,
-  rewardType,
-}) => {
-  if (paymentMethod) {
-    if (!newReward?.amount) {
-      setErrors({
-        ...errors,
-        tokenAmount: "Please enter the amount of tokens to be rewarded",
-      });
-      return;
-    }
-    // check if an existing reward is made with the same payment method - if so just edit
-    let existingReward = false;
-    const newRewards = rewards.map((reward) => {
-      if (reward.paymentMethodId === paymentMethod?.id) {
-        existingReward = true;
-        return {
-          ...reward,
-          amount: Number(newReward?.amount),
-        };
-      }
-      return reward;
-    });
-    if (existingReward) {
-      setQuestSettings((prev) => {
-        return {
-          ...prev,
-          rewards: newRewards,
-        };
-      });
-    } else {
-      onRewardAdd({
-        type: rewardType,
-        paymentMethodId: paymentMethod?.id,
-        paymentMethod,
-        amount: Number(newReward?.amount),
-      });
-    }
-    handleToggle();
-  } else if (addPaymentMethod) {
-    // Create payment method and then add reward
-    if (!newReward?.contractAddress) {
-      setErrors({
-        ...errors,
-        contractAddress: "Please enter the contract address",
-      });
-      return;
-    }
-    if (!newReward?.chain) {
-      setErrors({
-        ...errors,
-        chain: "Please select the appropriate chain",
-      });
-      return;
-    }
-    if (!newReward?.type) {
-      setErrors({
-        ...errors,
-        tokenType: "Please select the token type",
-      });
-      return;
-    }
-    if (!newReward?.amount) {
-      setErrors({
-        ...errors,
-        tokenAmount: "Please enter the amount of tokens to be rewarded",
-      });
-      return;
-    }
-    createPaymentMethod({
-      variables: {
-        input: {
-          orgId: activeOrg?.id,
-          contractAddress: newReward?.contractAddress,
-          tokenName: newReward?.tokenName,
-          symbol: newReward?.symbol,
-          icon: newReward?.icon,
-          chain: newReward?.chain,
-          type: newReward?.type.toUpperCase(),
-        },
-      },
-    })
-      .then((res) => {
-        const paymentMethod = res?.data?.createCmtyPaymentMethod;
-        onRewardAdd({
-          type: rewardType,
-          paymentMethodId: paymentMethod?.id,
-          amount: newReward?.amount,
-          paymentMethod,
-        });
-        handleToggle();
-        setAddPaymentMethod(false);
-      })
-      .catch((err) => {
-        setErrors({
-          ...errors,
-          paymentMethodCreation: err?.message,
-        });
-      });
-  }
 };
 
 const handleTokenRewardOnChange = ({ setQuestSettings, reward, value }) => {
@@ -282,122 +60,6 @@ const handleTokenRewardOnChange = ({ setQuestSettings, reward, value }) => {
   });
 };
 
-const handleAddPoap = ({ poapReward, setErrors, errors, onRewardAdd, rewardType }) => {
-  if (!poapReward?.id) {
-    setErrors({
-      ...errors,
-      poapEventId: "Please enter the POAP event ID",
-    });
-    return;
-  }
-  if (!poapReward?.name) {
-    setErrors({
-      ...errors,
-      poapEventId: "Please enter a valid POAP event ID",
-    });
-    return;
-  }
-  if (!poapReward?.eventSecret) {
-    setErrors({
-      ...errors,
-      eventSecret: "Please enter a poap event secret",
-    });
-    return;
-  }
-  onRewardAdd({
-    type: rewardType,
-    poapRewardData: poapReward,
-  });
-};
-
-const useAddRewardModalState = ({ paymentMethods }) => {
-  const [isRewardModalOpen, setIsRewardModalOpen] = useState(false);
-  const { isOpen: isTourOpen, setCurrentStep, currentStep, setSteps, steps } = useTour();
-  useEffect(() => {
-    if (isTourOpen) {
-      const newSteps = steps.map((step: any) => {
-        if (step.id === "tutorial-quest-rewards") {
-          return {
-            ...step,
-            handleNextAction: () => {
-              setIsRewardModalOpen(true);
-              setCurrentStep((prev) => prev + 1);
-            },
-          };
-        }
-        if (step.id === "tutorial-add-rewards") {
-          return {
-            ...step,
-            handleNextAction: () => {
-              setIsRewardModalOpen(false);
-              setCurrentStep((prev) => prev + 1);
-            },
-            handlePrevAction: () => {
-              setIsRewardModalOpen(false);
-              setCurrentStep((prev) => prev - 1);
-            },
-          };
-        }
-        if (step.id === "tutorial-activate-quest") {
-          return {
-            ...step,
-            handlePrevAction: () => {
-              setIsRewardModalOpen(true);
-              setCurrentStep((prev) => prev - 1);
-            },
-          };
-        }
-        return step;
-      });
-      setSteps(newSteps);
-    }
-  }, [isTourOpen]);
-  const [rewardType, setRewardType] = useState(PAYMENT_OPTIONS.DISCORD_ROLE);
-  const [discordRoleReward, setDiscordRoleReward] = useState(null);
-  const [paymentMethod, setPaymentMethod] = useState(null);
-  const [tokenReward, setTokenReward] = useState({
-    tokenName: null,
-    contractAddress: null,
-    symbol: null,
-    icon: null,
-    type: null,
-    chain: null,
-    amount: null,
-  });
-  const [editPaymentMethod, setEditPaymentMethod] = useState({
-    id: null,
-    tokenName: null,
-    contractAddress: null,
-    symbol: null,
-    icon: null,
-    type: null,
-    chain: null,
-    amount: null,
-  });
-  const [addPaymentMethod, setAddPaymentMethod] = useState(!paymentMethods.length);
-  const [poapReward, setPoapReward] = useState(null);
-  return {
-    isRewardModalOpen,
-    setIsRewardModalOpen,
-    rewardType,
-    setRewardType,
-    discordRoleReward,
-    setDiscordRoleReward,
-    paymentMethod,
-    setPaymentMethod,
-    tokenReward,
-    setTokenReward,
-    editPaymentMethod,
-    setEditPaymentMethod,
-    addPaymentMethod,
-    setAddPaymentMethod,
-    poapReward,
-    setPoapReward,
-    isTourOpen,
-    setCurrentStep,
-  };
-};
-
 const RewardComponent = ({
   rewards,
   setQuestSettings,
@@ -407,76 +69,23 @@ const RewardComponent = ({
   setQuestSettings: React.Dispatch<React.SetStateAction<{ [key: string]: any }>>;
   hasReferralStep: boolean;
 }) => {
-  const { activeOrg } = useContext(GlobalContext);
-  const { plan, setPaywall, setPaywallMessage } = useSubscriptionPaywall();
-  const { discordRoleOptions, discordRoleData } = useDiscordRoleRewardData();
-  const { createPaymentMethod, paymentMethodOptions, paymentMethods } = useTokenRewardData();
-  const [errors, setErrors] = useState(null);
   const onRewardAdd = (reward) => {
     setQuestSettings((prev) => {
+      const filteredRewards = prev.rewards.filter((i) => i.paymentMethodId !== reward.paymentMethod?.id);
       return {
         ...prev,
-        rewards: [...prev.rewards, reward],
+        rewards: [...filteredRewards, reward],
       };
     });
   };
-  const {
-    isRewardModalOpen,
-    setIsRewardModalOpen,
-    rewardType,
-    setRewardType,
-    discordRoleReward,
-    setDiscordRoleReward,
-    paymentMethod,
-    setPaymentMethod,
-    tokenReward,
-    setTokenReward,
-    editPaymentMethod,
-    setEditPaymentMethod,
-    addPaymentMethod,
-    setAddPaymentMethod,
-    poapReward,
-    setPoapReward,
-    isTourOpen,
-    setCurrentStep,
-  } = useAddRewardModalState({ paymentMethods });
+
+  const rewardModalState = useAddRewardModalState();
+  const { setIsRewardModalOpen, isTourOpen, setCurrentStep } = rewardModalState;
 
   const handleToggleModal = () => {
     setIsRewardModalOpen((prev) => !prev);
     if (isTourOpen) {
       setCurrentStep((prev) => prev + 1);
-    }
-  };
-
-  const handleAddRewardOnModal = () => {
-    if (rewardType === PAYMENT_OPTIONS.DISCORD_ROLE) {
-      handleNewDiscordRole({
-        newReward: { value: discordRoleReward, type: rewardType },
-        rewards,
-        onRewardAdd,
-        discordRoleData,
-        discordRoleOptions,
-      });
-      handleToggleModal();
-    } else if (rewardType === PAYMENT_OPTIONS.TOKEN) {
-      handleAddTokenOnModal({
-        newReward: tokenReward,
-        rewards,
-        onRewardAdd,
-        paymentMethod,
-        setErrors,
-        errors,
-        setQuestSettings,
-        handleToggle: handleToggleModal,
-        addPaymentMethod,
-        createPaymentMethod,
-        activeOrg,
-        setAddPaymentMethod,
-        rewardType,
-      });
-    } else if (rewardType === PAYMENT_OPTIONS.POAP) {
-      handleAddPoap({ poapReward, setErrors, errors, onRewardAdd, rewardType });
-      handleToggleModal();
     }
   };
 
@@ -538,99 +147,14 @@ const RewardComponent = ({
   const pointReward = rewards.filter((reward) => reward?.type === "points")[0];
   const otherRewards = rewards.filter((reward) => reward?.type !== "points");
 
-  const modalRewardButtonsProps = [
-    {
-      paymentOption: PAYMENT_OPTIONS.DISCORD_ROLE,
-      rewardType: rewardType,
-      onClick: () => setRewardType(PAYMENT_OPTIONS.DISCORD_ROLE),
-      Icon: DiscordRoleIcon,
-      text: "Discord Role",
-    },
-    {
-      paymentOption: PAYMENT_OPTIONS.POAP,
-      rewardType: rewardType,
-      onClick: () => setRewardType(PAYMENT_OPTIONS.POAP),
-      Icon: PoapIcon,
-      text: "POAP",
-    },
-    {
-      paymentOption: PAYMENT_OPTIONS.TOKEN,
-      rewardType: rewardType,
-      onClick: () => {
-        setRewardType(PAYMENT_OPTIONS.TOKEN);
-        if (plan === PricingOptionsTitle.Basic) {
-          setPaywall(true);
-          setPaywallMessage("This reward option is not available under the basic plan.");
-          setRewardType(PAYMENT_OPTIONS.DISCORD_ROLE);
-          return;
-        } else {
-          setRewardType(PAYMENT_OPTIONS.TOKEN);
-        }
-      },
-      Icon: TokensIcon,
-      text: "Token reward",
-    },
-  ];
-
   return (
     <>
-      <Modal
-        open={isRewardModalOpen}
-        onClose={handleToggleModal}
-        title="Add reward to quest"
-        modalComponentProps={{
-          className: "tour-default-modal",
-        }}
-        dialogComponentProps={{
-          className: "tutorials-quest-reward-modal",
-        }}
-        maxWidth={800}
-        footerLeft={
-          <RewardFooterLeftComponent
-            rewardType={rewardType}
-            paymentMethod={paymentMethod}
-            setPaymentMethod={setPaymentMethod}
-            addPaymentMethod={addPaymentMethod}
-            handleReward={handleAddRewardOnModal}
-            setAddPaymentMethod={setAddPaymentMethod}
-            editPaymentMethod={editPaymentMethod}
-            setEditPaymentMethod={setEditPaymentMethod}
-          />
-        }
-        footerRight={undefined}
-        footerCenter={undefined}
-      >
-        <Grid display="flex" flexDirection="column" gap="24px">
-          <Grid container item gap="14px">
-            <Label>Reward Type</Label>
-            <Grid container item alignItems="center" gap="14px" width="100%" justifyContent="center">
-              {modalRewardButtonsProps.map((props) => (
-                <RewardMethodOptionButton {...props} />
-              ))}
-            </Grid>
-          </Grid>
-          <Grid container item flexDirection="column" gap="14px">
-            <RewardMethod
-              rewardType={rewardType}
-              componentsOptions={discordRoleOptions}
-              discordRoleReward={discordRoleReward}
-              setDiscordRoleReward={setDiscordRoleReward}
-              tokenReward={tokenReward}
-              setTokenReward={setTokenReward}
-              paymentMethod={paymentMethod}
-              paymentMethods={paymentMethods}
-              addPaymentMethod={addPaymentMethod}
-              setPaymentMethod={setPaymentMethod}
-              editPaymentMethod={editPaymentMethod}
-              setEditPaymentMethod={setEditPaymentMethod}
-              errors={errors}
-              setErrors={setErrors}
-              poapReward={poapReward}
-              setPoapReward={setPoapReward}
-            />
-          </Grid>
-        </Grid>
-      </Modal>
+      <RewardModal
+        rewardModalState={rewardModalState}
+        handleRewardModalToggle={handleToggleModal}
+        handleOnRewardAdd={onRewardAdd}
+        rewards={rewards}
+      />
 
       <Grid container flexWrap="nowrap" maxWidth="inherit">
         <Grid container item gap="14px" alignItems="center" justifyContent="center" flex="1">
